+++ src/main/java/com/green/smartgradever2/student/model/StudentLectureReg.java
@@ -7,5 +7,5 @@ import lombok.Setter;
 @Setter
 public class StudentLectureReg {
     private Long ilecture;
-    private Integer studentNum;
+    private Long studentNum;
 }
